// modules/diplom-projects/src/app/models/task-sheet.model.ts
export class TaskSheet {
  DiplomProjectId: string;
  InputData: string;
  Faculty: string;
  HeadCathedra: string;
  RpzContent: string;
  DrawMaterials: string;
  Univer: string;
  DateEnd: string;
  DateEndString: string;
  DateStart: string;
  DateStartString: string;
  ComputerConsultant: string;
  HealthAndSafetyConsultant: string;
  EconimicsConsultant: string;
  NormocontrolConsultant: string;
  DecreeNumber: string;
  DecreeDate: string;
  DecreeDateString: string;
}